% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country256-package.R
\docType{package}
\name{country256-package}
\alias{country256}
\alias{country256-package}
\title{country256: Encode Countries as Raw Vectors}
\description{
Given there are fewer than 256 countries in typical formulations, encoding them as a vector of 1-byte elements can conserve memory.
}
\author{
\strong{Maintainer}: Hugh Parsonage \email{hugh.parsonage@gmail.com}

}
\keyword{internal}
