% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_country.R
\name{encode_country256}
\alias{encode_country256}
\alias{decode_country256}
\title{Encode country as 8-bit}
\usage{
encode_country256(x, na_treatment = "__auto__")

decode_country256(x, to = "iso3", na_treatment = "__auto__")
}
\arguments{
\item{x}{A character vector to encode/decode.}

\item{na_treatment}{A method of treating NA values.}

\item{to}{Either \code{"iso2"} or \code{"iso3"}, designating the code
the encoded values should be decoded to. Essentially, whether to decode to
2-character or 3-character strings.}
}
\value{
For encode, a raw vector; for decode, a (standardized) character vector.
}
\description{
Encode country as 8-bit
}
\details{
Since \code{"NA"} is the correct 2-character ISO-3166 encoding for Namibia,
such strings should not necessarily be interpreted as \code{NA_character_},
yet in some contexts, this may be the appropriate choice.  The parameter
\code{na_treatment} is concerned with this choice. In this version,
\code{"__auto__"} treats any such string as Namibia if all the other strings
are length-2
}
